refactor: rename TableFooter and TableHead components to Footer and Head for consistency
