+++ src/codeblocks/ui/componets/Table/Footer/Footer.svelte
@@ -4,7 +4,7 @@
 <tfoot>
   <tr>
     <td></td>
-    <td> Footer </td>
+    <td>Footer</td>
     <td></td>
     <td></td>
   </tr>
